requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.7.0
trl>=0.7.0
unsloth>=2024.9
vllm>=0.2.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl  
tqdm>=4.65.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
codecarbon>=2.3.0
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
